% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-io.R
\name{loadFile}
\alias{loadFile}
\title{load file data and return it as vector of character strings}
\usage{
loadFile(file, row.type)
}
\description{
load file data and return it as vector of character strings
}
