% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-io.R
\name{loadTableName}
\alias{loadTableName}
\title{Get CSA Table name from file}
\usage{
loadTableName(file)
}
\value{
Character value that represents OFSAA CSA table name \code{STG_...}
}
\description{
Function takes filename and returns corresponding table name.
  Table name is stored in row_type = 0 in position 4
}
