% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{processFiles}
\alias{processFiles}
\title{Process files}
\usage{
processFiles(input.folder = ".", file.names = "*", output.folder = "",
  rules.file = "", seed = 0)
}
\arguments{
\item{input.folder}{- input folder, word directory by default}

\item{file.names}{- file wildcard to select files}

\item{output.folder}{-}
}
\description{
This procedure scrambles all files which meet selection criteria according to scramble rules
}
