% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-io.R
\name{saveFile}
\alias{saveFile}
\title{Write data into file
function takes parts of file and writes it in OFSAA compatible format}
\usage{
saveFile(header, columns, data, footer, file)
}
\description{
Write data into file
function takes parts of file and writes it in OFSAA compatible format
}
