% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scramble.R
\name{scramble}
\alias{scramble}
\title{Scramble a dataset}
\usage{
scramble(data, check.identical = TRUE)
}
\arguments{
\item{data}{Data.frame. The original data. No default.}

\item{check.identical}{Logical. If TRUE \code{scramble} checks that no
observation in the original dataset is present in the scrambled
dataset. The probability that this would happen is of course
minimal, but this check is here as an fail-safe. Note that this
operation may take a long time with large datasets, and can
therefore be turned off at your own risk. Defaults to TRUE.}
}
\description{
This is the workhorse function of the \code{scramble} package. It takes
a dataset as input and returns a scrambled dataset. The scrambled
dataset includes the same columns as the original dataset but these
have been resampled in such a way that the associations between
columns present in the original dataset are lost.
}
