% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrambe.R
\name{scrambleDataFrame}
\alias{scrambleDataFrame}
\title{Scramble dataframe values}
\usage{
scrambleDataFrame(data, tablename, seed = 100, scrambling.rules)
}
\arguments{
\item{data}{- dataframe with data to be scrambles}

\item{tablename}{- tablename (reference to table in scrambling rules)}

\item{seed}{- seed value for randomization}

\item{scrambling.rules}{- scrambling rules}
}
\value{
Function returns data frame with scrambled values.
}
\description{
The function scrambles vector or values based on method provided
}
