% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrambe.R
\name{scrambleValue}
\alias{scrambleValue}
\title{Scramble values}
\usage{
scrambleValue(value, method, seed = 100, new.value = "", max.len = "")
}
\arguments{
\item{value}{- vector of values to scramble}

\item{method}{- obfuscation bethod. Supported methods are \code{shuffle},
\code{hash}, \code{random.hash}, \code{random.num}, \code{rnorm.num}}

\item{seed}{- seed value for random generation and sampling}

\item{new.value}{- fixed value to replace original one (works only with \code{fixed.value} method)}

\item{max.len}{- maximum length of scrabled value (useful when data column is of limited length)}
}
\description{
The function scrambles vector or values based on method provided
}
