% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scrambling.rules}
\alias{scrambling.rules}
\title{Rules for data scrambling}
\format{A dataframe with 5 variables
\describe{
  \item{File}{
    File name / regular expression - used to store different rules for different files
  }
  \item{Column}{
    dataframe column
  }
  \item{Method}{
    scrambling method; possible values are described in
    \code{method} parameter of \code{scrambleValue} function
  }
  \item{Fixed.Value}{
    use fixed value to replace original values - works together with
    \code{fixed.value} method
  }
  \item{Max.Length}{
    maximum length of scrambled value
  }
}}
\usage{
scrambling.rules
}
\description{
A dataset contains rules that a subsequently applied to data-frame columns.
Each rule scrambles column using set of package functions.
}
\keyword{datasets}
