% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrambe.R
\name{scrambleDataFrame}
\alias{scrambleDataFrame}
\title{Scramble dataframe values}
\usage{
scrambleDataFrame(data, tablename, seed = 100, scramble.rules)
}
\description{
The function scrambles vector or values based on method provided
}
