% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrambe.R
\name{scrambleValue}
\alias{scrambleValue}
\title{Scramble values}
\usage{
scrambleValue(value, method, seed = 100, method.param = "", max.len)
}
\arguments{
\item{value}{- vector of values to scramble}

\item{method}{- obfuscation bethod. Supported methods are \code{shuffle},
\code{hash}, \code{random.hash}, \code{random.num}, \code{rnorm.num}}

\item{seed}{- seed value for random generation and sampling}

\item{method.param}{- additional information associated with method; for example hash algorithm or exact fixed value}

\item{max.len}{- maximum length of scrabled value (useful when data column is of limited length)}
}
\description{
The function scrambles vector or values based on method provided
}
