% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrambe.R
\name{scrambleValue}
\alias{scrambleValue}
\title{Scramble values}
\usage{
scrambleValue(value, method, seed = 100, new.value = "", max.len)
}
\arguments{
\item{value}{- vector of values to scramble}

\item{method}{- obfuscation bethod. Supported methods are \code{shuffle},
\code{hash}, \code{random.hash}, \code{random.num}, \code{rnorm.num}}
}
\description{
The function scrambles vector or values based on method provided
}
